Provide formatting/settings to help contributors
Reading the contributions guidelines is far less effective
than providing automated tooling that can deal with the
little big details like tabs and braces :).

Coupled with the EditorConfig extension (which also works
on VSCode and a myriad other text editor) as well as one
of the many extensions in VS that automatically load
.vssettings files alongside the solution (which can also
be imported manually of course), this ensures that a
contributor knows right-away what formatting to use,
without having to think.

I've adjusted by VS settings to the observed coding
conventions (which are pretty much the VS defaults,
but not Xamarin/Mono's defaults, for example ;)), and
exported the matching settings file. Likewise, I've
set up the .editorconfig with the spacing used in the
commonly edited files I found in the solution.